Use random tempfile for `LYCHEE_TMP`
This prevents output overlapping when running lychee multiple times.

Note that the final output file remains unchanged.
This is on purpose because when executing lychee multiple times
within the same workflow, the output of all runs will be concatenated.
To prevent that you can set a different output path for both runs:

```yaml
      - uses: lycheeverse/lychee-action@master
        with:
          args: '<some args> -- <file1>'
          output: run1.md
          fail: true
          jobSummary: true

      - uses: lycheeverse/lychee-action@master
        with:
          args: '<some args> -- <file2>'
          output: run2.md
          fail: true
          jobSummary: true
```