Work around new naming scheme
Until version 0.15.1, the naming scheme of the release assets let them
start with `lychee-<version>-`. This is no longer the case as of
https://github.com/lycheeverse/lychee/pull/1464, though, breaking the
`lychee-action`.

Since we need to allow using older versions, still, we are now stuck
with a really ugly `case` construct, but this is still far better than
having a broken GitHub Action.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>