Adjust for new asset naming scheme (#234)
Work around new naming scheme

Until version 0.15.1, the naming scheme of the release assets let them
start with `lychee-<version>-`. This is no longer the case as of
https://github.com/lycheeverse/lychee/pull/1464, though, breaking the
`lychee-action`.

Since we need to allow using older versions, still, we are now stuck
with a really ugly `case` construct, but this is still far better than
having a broken GitHub Action.

The name of the `.tar` file is repeated multiple times, verbatim. This
not only violates the Don't Repeat Yourself (DRY) principle, it also
makes it easier to miss an instance when the file name changes (as it
did with https://github.com/lycheeverse/lychee/pull/1464).
DRY up the lengthy `.tar` file name.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>